refactor[GetDocumento]
