+++ src/modules/GetDocumento/dtos/IGetDocumentoDTO.ts
